Refactor: split Vocabularies parser in multiple files
